Important fix to OrderOf
Simply comparing the labels (the notation `v(x)` in Dietz/Sleator) in
OrderOf yields incorrect results if the base record has changed from its
initial state. Reading the paper more carefully, I noticed that OrderOf
uses a comparison of the offsets of the labels from the label of the
base record (the notation `v sub b (x)` in Dietz/Sleator). Making this
change and using OrderOf  in the tests allows them to succeed at
maintaining order.