Various fixes/additions to OrderKeeper
AddBefore() is new, and does what you'd expect. An OrderKeeper can be
used as an iterator over its T values in order from first to last, which
does not necessarily coincide with the earlier index-based ordering
(which is handled by KeyedCollection, and is insertion-ordered). Various
fixes were found and implemented while working on AddBefore, and this
can get to the point where there isn't enough RAM to continue adding
items, without hitting any duplicate labels or order mismatch. I think
it should be good for usage.