Changes to Herd, add PRNG class for speed
HerdRandomness is a little faster now and has better quality in some
ways; this commit brings the algorithm up to date with the Java library
sarong, which is being developed in tandem. PRNG is effectively RNG with
a hard-coded usage of HerdRandomness instead of a swappable Randomness,
and it is quite a bit faster than RNG by default both on 64-bit and
32-bit targets, while keeping the high period and quality of Herd. PRNG
is, like RNG, a subclass of System.Random, and PRNG should be a good
general-purpose replacement.