ThrustRandomness/PRNG3 is now fastest
A Java version with an identical algorithm and constants (though using
Java's signed `int` and `>>>` instead of C#'s `uint` and `>>`) passed
PractRand testing with no statistical failures or anomalies, and the
version in PRNG3 is now the fastest one we have (roughly 4x as fast as
System.Random, but with a short period of 2 to the 32).

The benchmarks have been updated to use slightly more professional
techniques.