Add Lathe and Linnorm
Lathe is recommended when you want a smallish state (it fits in one ulong) and still high quality/speed on 32-bit outputs, on both x86 and x64 runtimes. It's the fastest way to produce int or uint output here that still passes quality tests.

Linnorm has excellent speed on the JVM, but is sorta not as good here, even with x64 .NET as the runtime. It's still good, and its Determine() method should probably be copied into the other implementations since it's very fast as stateless random methods go, and still passes quality tests.