Change all NextLong(long)
These now use a much faster technique that fakes a 128-bit multiplication of a long (or ulong) bound with a random ulong and uses the high 64 bits of that product. This is biased, but for a long enough period (2 to the 64 should be fine) detecting any bias is rather infeasible.